LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/css-module/script-element-css-src-expected.txt

CONSOLE MESSAGE: TypeError: 'text/css' is not a valid JavaScript MIME type.

PASS Test that <script> doesn't load when the src is CSS.
